% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smap.R
\name{plot_biomes_map}
\alias{plot_biomes_map}
\title{Plot Custom-Colored Biomes Map}
\usage{
plot_biomes_map(
  biomes,
  biome_colors,
  draw_countries = FALSE,
  draw_coastline = TRUE,
  projection = "+proj=robin"
)
}
\arguments{
\item{biomes}{An 'sf' object representing the biome data, expected to include a
\code{BIOME_NAME} column which matches the keys in \code{biome_colors}.}

\item{biome_colors}{A named vector providing colors for each biome, where the names
must correlate with the unique \code{BIOME_NAME}s in the \code{biomes} dataset.}

\item{draw_countries}{Logical indicating if country borders should be overlaid on the map.
Uses 'rnaturalearth'. Defaults to \code{FALSE}.}

\item{draw_coastline}{Logical indicating if the global coastline should be drawn.
Defaults to \code{TRUE}.}

\item{projection}{Map projection parameters in PROJ format.
Defaults to Robinson projection (\code{"+proj=robin"}).}
}
\value{
A \code{ggplot} object depicting the biomes map which can be either directly displayed
using \code{print()} or further customized.
}
\description{
Visualizes a global map of biomes with custom color coding, with options
for overlaying country borders and coastlines. This function depends on the 'sf',
'ggplot2', and 'rnaturalearth' packages to appropriately handle spatial data and mapping.
The biome data is expected to be loaded with \code{data("biomes")}, which should produce
an 'sf' object including a column for biome names (\code{BIOME_NAME}).
}
\examples{
# biomes data can be downloaded from https://ecoregions.appspot.com/
library(sf)
biomes <- st_read("Ecoregions2017.shp")

# Define custom colors for biomes
biome_colors <- c(
  "Boreal Forests/Taiga" = "#7AB6F5",
  "Deserts & Xeric Shrublands" = "#CC6767",
  "Flooded Grasslands & Savannas" = "#BEE7FF",
  "Mangroves" = "#FE01C4",
  "Mediterranean Forests, Woodlands & Scrub" = "#FE0000",
  "Montane Grasslands & Shrublands" = "#D6C39D",
  "Temperate Broadleaf & Mixed Forests" = "#00734C",
  "Temperate Conifer Forests" = "#458970",
  "Temperate Grasslands, Savannas & Shrublands" = "#FEFF73",
  "Tropical & Subtropical Coniferous Forests" = "#88CE66",
  "Tropical & Subtropical Dry Broadleaf Forests" = "#CCCD65",
  "Tropical & Subtropical Grasslands, Savannas & Shrublands" = "#FEAA01",
  "Tropical & Subtropical Moist Broadleaf Forests" = "#38A700",
  "Tundra" = "#9ED7C2"
)

# Create and display the biomes map, without country borders and with coastline
plot_biomes_map(biomes, biome_colors, draw_countries = FALSE, draw_coastline = TRUE, projection = "+proj=robin")

}
