% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smap.R
\name{plot_map}
\alias{plot_map}
\title{Plot World Map}
\usage{
plot_map(
  draw_coastline = FALSE,
  fill_color = "#e0e0eb",
  projection = "+proj=robin"
)
}
\arguments{
\item{draw_coastline}{Logical flag to indicate whether to draw the coastline.
Default is \code{TRUE}.}

\item{fill_color}{The hexadecimal code for the countries' fill color.
Default is \code{"#e0e0eb"}, a light grey-blue.}

\item{projection}{Cartographic projection in PROJ format.
Default is \code{"'+proj=robin'"} for Robinson projection.}
}
\value{
A \code{ggplot} object representing the world map that can be either further customized
with 'ggplot2' syntax or displayed using the \code{print()} function.
}
\description{
This function creates a simple, customizable world map. Options include toggling
drawing coastlines, setting fill color, and choosing map projection. It relies
on 'ggplot2' for plotting, 'rnaturalearth' for geographic data, and 'sf' for spatial
operations.
}
\examples{

# Plot a simple world map with default settings:
plot_map()

# Plot a world map showing with coastline
plot_map(draw_coastline = TRUE,fill_color = "#A6D96A")

}
