\name{SMAP-internal}
\docType{methods}
\alias{A,grad-method}
\alias{A,eta-method}
\alias{Phi,grad-method}
\alias{Phi,eta-method}
\alias{Pi,grad-method}
\alias{Pi,eta-method}
\alias{initialize,grad-method}
\alias{initialize,eta-method}
\alias{initialize,gaussparam-method}
\alias{value,eta-method}
\alias{P,SMAPProfile-method}
\alias{Q,SMAPProfile-method}
\alias{Q,SMAPProfiles-method}
\alias{HMM,SMAPProfile-method}
\alias{initialize,SMAPProfile-method}
\alias{observations,SMAPProfile-method}
\alias{observations,SMAPProfiles-method}
\alias{show,SMAPProfile-method}
\alias{show,SMAPProfiles-method}
\alias{name,SMAPProfile-method}
\alias{name,SMAPProfiles-method}
\alias{[,SMAPProfile-method}
\alias{[,SMAPObservations-method}
\alias{[[<-,SMAPProfiles-method}
\alias{chromosome,SMAPObservations-method}
\alias{endPosition,SMAPObservations-method}
\alias{initialize,SMAPObservations-method}
\alias{name,SMAPObservations-method}
\alias{noObservations,SMAPObservations-method}
\alias{reporterId,SMAPObservations-method}
\alias{show,SMAPObservations-method}
\alias{startPosition,SMAPObservations-method}
\alias{value,SMAPObservations-method}
\alias{A,SMAPHMM-method}
\alias{Phi,SMAPHMM-method}
\alias{Pi,SMAPHMM-method}
\alias{initialize,SMAPHMM-method}
\alias{noStates,SMAPHMM-method}
\alias{show,SMAPHMM-method}
\alias{smap,SMAPHMM,SMAPObservations-method}
\alias{plot,SMAPProfile,missing-method}
\alias{plot,SMAPProfiles,missing-method}
\title{SMAP-internal}
\description{Internal methods for the \pkg{SMAP} package.}
\seealso{
  \code{\link[SMAP:smap]{smap}}
}
\author{Robin Andersson, \email{robin.andersson@lcb.uu.se}}
\references{
 Andersson, R., Bruder, C. E. G., Piotrowski, A., Menzel, U., Nord, H.,
  Sandgren, J., Hvidsten, T. R., Diaz de Stahl, T., Dumanski, J. P.,
Komorowski, J., A Segmental Maximum A Posteriori Approach to Array-CGH
Copy Number Profiling, submitted
}
\keyword{methods}
\keyword{internal}
